Create Default DSC
