+++ pkg/plugins/addLabelsplugin.go
@@ -10,20 +10,12 @@ import (
 func ApplyAddLabelsPlugin(componentName string, resMap resmap.ResMap) error {
 	nsplug := builtins.LabelTransformerPlugin{
 		Labels: map[string]string{
-			"app.kubernetes.io/part-of":           componentName,
 			"app.opendatahub.io/" + componentName: "true",
 		},
 		FieldSpecs: []types.FieldSpec{
 			{
 				Gvk:                resid.Gvk{Kind: "Deployment"},
 				Path:               "spec/template/metadata/labels",
-				CreateIfNotPresent: true,
-			},
-			{
-				Gvk: resid.Gvk{
-					Kind: "Deployment",
-				},
-				Path:               "spec/selector/matchLabels",
 				CreateIfNotPresent: true,
 			},
 			{
